refactor(api): isolate deployment state machine
