+++ apps/api/src/repositories/deployment.repository.ts
@@ -1,8 +1,4 @@
 import type { DeploymentId, DeploymentState } from '@repo/protocol';
-import * as Data from 'effect/Data';
-import * as Effect from 'effect/Effect';
-import * as Layer from 'effect/Layer';
-import * as Option from 'effect/Option';
 import type { Queries } from '#db/queries.gen.d.ts';
 import { Repository } from '#repositories/repository.ts';
 
@@ -20,29 +16,6 @@ export type DeploymentTransition = {
   readonly expectedState: DeploymentState;
   readonly next: Omit<DeploymentStateSnapshot, 'id'>;
 };
-
-type DeploymentStateStoreService = {
-  find(input: {
-    deploymentId: DeploymentId;
-  }): Effect.Effect<Option.Option<DeploymentStateSnapshot>, DeploymentPersistenceError>;
-  findExpired(input: {
-    asOf: Date;
-    limit: number;
-  }): Effect.Effect<readonly DeploymentStateSnapshot[], DeploymentPersistenceError>;
-  transition(
-    input: DeploymentTransition,
-  ): Effect.Effect<Option.Option<DeploymentStateSnapshot>, DeploymentPersistenceError>;
-};
-
-export class DeploymentPersistenceError extends Data.TaggedError('DeploymentPersistenceError')<{
-  operation: keyof DeploymentStateStoreService;
-  cause: unknown;
-}> {}
-
-export class DeploymentStateStore extends Effect.Tag('#repositories/deployment-state-store')<
-  DeploymentStateStore,
-  DeploymentStateStoreService
->() {}
 
 type DeploymentStateRow = Queries['SelectDeploymentState'];
 type ExpiredDeploymentStateRow = Queries['SelectExpiredDeploymentStates'];
@@ -138,44 +111,4 @@ export class DeploymentRepository extends Repository {
     const row = rows[0];
     return row ? toSnapshot(row) : null;
   }
-}
-
-function persistenceEffect<Value>({
-  operation,
-  execute,
-}: {
-  operation: DeploymentPersistenceError['operation'];
-  execute: () => Promise<Value>;
-}): Effect.Effect<Value, DeploymentPersistenceError> {
-  return Effect.tryPromise({
-    try: execute,
-    catch: (cause) => new DeploymentPersistenceError({ operation, cause }),
-  });
-}
-
-export function deploymentStateStoreLayer({
-  repository,
-}: {
-  repository: DeploymentRepository;
-}): Layer.Layer<DeploymentStateStore> {
-  return Layer.succeed(DeploymentStateStore, {
-    find({ deploymentId }) {
-      return persistenceEffect({
-        operation: 'find',
-        execute: async () => Option.fromNullable(await repository.find({ deploymentId })),
-      });
-    },
-    findExpired({ asOf, limit }) {
-      return persistenceEffect({
-        operation: 'findExpired',
-        execute: async () => await repository.findExpired({ asOf, limit }),
-      });
-    },
-    transition(input) {
-      return persistenceEffect({
-        operation: 'transition',
-        execute: async () => Option.fromNullable(await repository.transition(input)),
-      });
-    },
-  });
 }
